+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/service/message/PersonDeletionService.kt
@@ -5,22 +5,19 @@ import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
 import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.PersonEntity
 import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.PersonKeyEntity
-import uk.gov.justice.digital.hmpps.personrecord.jpa.repository.PersonKeyRepository
 import uk.gov.justice.digital.hmpps.personrecord.jpa.repository.PersonRepository
-import uk.gov.justice.digital.hmpps.personrecord.jpa.repository.ReviewRepository
 import uk.gov.justice.digital.hmpps.personrecord.service.cprdomainevents.events.person.PersonDeleted
-import uk.gov.justice.digital.hmpps.personrecord.service.cprdomainevents.events.personkey.PersonKeyDeleted
 import uk.gov.justice.digital.hmpps.personrecord.service.message.recluster.ReclusterService
+import uk.gov.justice.digital.hmpps.personrecord.service.person.PersonKeyDeletionService
 import uk.gov.justice.digital.hmpps.personrecord.service.search.PersonMatchService
 
 @Component
 class PersonDeletionService(
   private val personRepository: PersonRepository,
-  private val personKeyRepository: PersonKeyRepository,
   private val personMatchService: PersonMatchService,
   private val publisher: ApplicationEventPublisher,
   private val reclusterService: ReclusterService,
-  private val reviewRepository: ReviewRepository,
+  private val personKeyDeletionService: PersonKeyDeletionService,
 ) {
 
   @Transactional
@@ -49,12 +46,7 @@ class PersonDeletionService(
   private fun PersonEntity.deleteFromPersonMatch() = personMatchService.deleteFromPersonMatch(this)
 
   private fun deletePersonKey(personKeyEntity: PersonKeyEntity, personEntity: PersonEntity) {
-    val findByClustersPersonKey = reviewRepository.findByClustersPersonKey(personKeyEntity)
-    findByClustersPersonKey?.let { review ->
-      reviewRepository.delete(review)
-    }
-    personKeyRepository.delete(personKeyEntity)
-    publisher.publishEvent(PersonKeyDeleted(personEntity, personKeyEntity))
+    personKeyDeletionService.deletePersonKey(personKeyEntity, personEntity)
   }
 
   private fun PersonEntity.deletePersonEntityThatWasMergedIntoThisOneRecursively() {
